# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dashboard Pimpinan",
        "description": "",
        "contact": {
            "name": "Developer",
            "email": "diskominfo@bandung.go.id"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/add-master-data-wifi": {
            "post": {
                "tags": [
                    "Master Data Wifi"
                ],
                "summary": "Master Data Wifi",
                "description": "/api/v1/add-master-data-wifi",
                "operationId": "Master Data Wifi",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "lokasi",
                                    "latitude",
                                    "longitude",
                                    "status"
                                ],
                                "properties": {
                                    "lokasi": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "vendor": {
                                        "type": "string"
                                    },
                                    "dinas": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Input Data Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Input Data Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}